+++ previous-versions/accounting/pkg/models/operations/downloadpurchaseorderpdf.go
@@ -3,6 +3,7 @@
 package operations
 
 import (
+	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/models/shared"
 	"io"
 	"net/http"
 )
@@ -34,6 +35,8 @@ type DownloadPurchaseOrderPdfResponse struct {
 	// Success
 	// The Close method must be called on this field, even if it is not used, to prevent resource leaks.
 	Data io.ReadCloser
+	// Your API request was not properly authorized.
+	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -52,6 +55,13 @@ func (o *DownloadPurchaseOrderPdfResponse) GetData() io.ReadCloser {
 		return nil
 	}
 	return o.Data
+}
+
+func (o *DownloadPurchaseOrderPdfResponse) GetErrorMessage() *shared.ErrorMessage {
+	if o == nil {
+		return nil
+	}
+	return o.ErrorMessage
 }
 
 func (o *DownloadPurchaseOrderPdfResponse) GetStatusCode() int {