+++ previous-versions/accounting/pkg/models/operations/uploadtransferattachment.go
@@ -67,7 +67,7 @@ func (o *UploadTransferAttachmentRequest) GetTransferID() string {
 type UploadTransferAttachmentResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
+	// The request made is not valid.
 	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int