+++ previous-versions/accounting/RELEASES.md
@@ -391,3 +391,13 @@ Based on:
 - [go v0.28.0] previous-versions/accounting
 ### Releases
 - [Go v0.28.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.28.0 - previous-versions/accounting
+
+## 2023-10-30 01:28:38
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.28.1] previous-versions/accounting
+### Releases
+- [Go v0.28.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.28.1 - previous-versions/accounting